created and started NSFRB injection test and implemented dask queue for cand cutting
