+++ nsfrb/noise.py
@@ -4,9 +4,10 @@ import sys
 import os
 import glob
 
-f = open("../metadata.txt","r")
-cwd = f.read()[:-1]
-f.close()
+#f = open("../metadata.txt","r")
+#cwd = f.read()[:-1]
+#f.close()
+cwd = os.environ['NSFRBDIR']
 sys.path.append(cwd + "/")
 from nsfrb.config import *
