+++ nsfrb/outputlogging.py
@@ -1,7 +1,9 @@
 #simple function to print to logging file during execution
-f = open("../metadata.txt","r")
-cwd = f.read()[:-1]
-f.close()
+#f = open("../metadata.txt","r")
+#cwd = f.read()[:-1]
+#f.close()
+import os
+cwd = os.environ['NSFRBDIR']
 
 output_file = cwd + "-logfiles/run_log.txt"#"/home/ubuntu/proj/dsa110-shell/dsa110-nsfrb/tmpoutput/run_log.txt"
