+++ process_server/run_proc_server
@@ -2,4 +2,4 @@
 
 #run this script to open server and receive 32x32x25x16 image data
 > $PWD/process_log.txt #/home/ubuntu/proj/dsa110-shell/dsa110-nsfrb/process_server/process_log.txt 
-python run.py "$@" --spacefilter --usefft --samenoise --cuda --initframes --initnoise --appendframe --usejax 
+python run.py --spacefilter --usefft --samenoise --cuda --initframes --initnoise --appendframe --usejax "$@"
